requirements.txt
google-genai
python-dotenv
black
matplotlib
PyQt6
pillow